pyproject.toml
[project]
name = "ocr-service"
version = "0.1.0"
description = "Servicio de OCR y procesamiento de documentos"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.136.1",
    "uvicorn[standard]==0.46.0",
    "sqlalchemy==2.0.49",
    "psycopg2-binary==2.9.12",
    "pydantic==2.13.3",
    "pydantic-settings==2.14.0",
    "python-multipart==0.0.27",
    "python-dotenv==1.2.2",
    "azure-ai-documentintelligence==1.0.2",
    "azure-core==1.39.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
